@@ -1,9 +1,6 @@
-import Vue from "vue";
 import Vuex from "vuex";
 
-Vue.use(Vuex);
-
-export default new Vuex.Store({
+export default Vuex.createStore({
   state: {},
   mutations: {},
   actions: {},
